Allow setting cover art via DnD and context menu
You can drag images onto the bottom-left cover art (if shown), or the
image shown in the toolbar, or to the image shown in the track metadata
editor.  You can also right click on any of those to delete the current
image or choose a new one via a file dialog.  Fixes BGO #336350