[build] Support building on OS X 10.6, drop 10.4
Dropped OS X 10.4 support, and now build against the 10.5 SDK.
Additionally we are building 32 bit binaries since Mono does
not support 64 bit on OS X. 64 bit is the default now on 10.6.

Updated the dependencies to the latest released versions. The
two build patches we had against GStreamer are now removed.

Bootstrapping gettext is no longer required, so it is removed.

Going forward we will support OS X 10.5 and 10.6.