[TrackEditor] Make sync-all sync the track count
Add an optional SyncClosure that can be used instead of the WriteClosure
for sync (eg so we can only sync the track count, not track number).
Add FieldOptions.NoShowSync for use by track number, since it already
has it's auto-set button.  Change the dialog to enumerate the syncable
fields by asking each FieldPage for its fields, instead of looking for
sync buttons in each notebook page.