[build] new more flexible dependency build system
Implemented a more flexible and simple system for building dependencies
that will be able to more easily build Mono and GTK, which is not as
straight forward as the other dependencies. Also a lot easier to maintain
and update the dep stack.

Should also be possible to extend the OS X profile to build a true bundle
for Linux as well. Would be an interesting hack.