[build] bundling now supported in build profile
The --bundle option can be passed to a build profile to create
a bundle from the installed build root. Currently most of the
Darwin/OS X bundle is implemented, with a few things left to do.

Define a bundle profile for Banshee in profile.darwin.py, and
added .app skeleton for Banshee.