[osx] store bundle path, relocate once, spaceguard
Only generate config files that need absolute paths when it is
detected that the bundle location has changed. Store the bundle
location in ~/.config/banshee-1/app-bundle-location. Additionally
remove the Mono Addins databases when we are relocated.

Finally, until a better work around for Pango modules is found,
the bundle path cannot have any spaces in it, since the resulting
absolute path as written to pango.modules will contain a space,
and pango will not properly parse this, so no modules will load
and Banshee will not be able to render any text.

If a space is detected in the bundle path, show a lame dialog
telling the user how lame this is.