[grid] massively improve rendering performance
While I have always oh so carefully computed proper clipping/damage
rectangles which prevent unnecessary regions of the view being
repainted on screen, I have never used this data to prevent walking
the scene unnecessarily.

That is, if we're not going to paint something on screen, don't
go through the motions to lay it out in memory either.

Face->Palm. Yet another thing that became more obvious and simple
by moving to this awesome new layout system.

I need to go for a beer run. Also, Sandy is a great listener.