[dap] Make DAP management less frustrating
Always allow manually managing it, even if a sync option is selected; if
you later press Sync, your changes will of course be overwritten, but
this gives you the ability to tweak the sync results.  Also, show the
Sync action even if auto-sync is on, in case it's not perfect at
detecting when to sync.