[libbanshee] Only pass playbin's state-changed messages to managed code
Previously, all state-changed messages on the bus would be passed to the managed state_changed_cb.  The managed code only cares about the state of the playbin; it doesn't need to know when each individual element in the pipeline changes state.
This change is correct for trunk, but is more important for the gapless branch, where the extra state-changed messages confuse things.