[Equalizer] Big refactor of EQ preset management
As I began to merge some predefined presets into the equalizer, I ran into
quite a few confusing, inefficient, and redundant chunks of code, scattered
across various places leading to lots of confusion.

This refactor better defines the roles of EqualizerManager vs EqualizerSetting.

In addition, I changed from an XML storage format to JSON. After moving from
the slow XmlDocument (DOM-based) API to XmlTextReader (tokenizer API), which
saved lots of time loading the presets, I then tried the Hyena JSON parser,
which saved even more time. The progression went something like this:

  XML DOM (0.075s) => XML Tokenizer (0.012s) => Hyena JSON (0.007s)

The old XML format will be loaded if the JSON format is not found, and in turn
saved to the new JSON format, so no presets should be lost.

Additionally, due to the clean up, equalizer band gain values should not be
excessively flushed to the engine, which could solve some of the noise-related
bugs that have been filed (not confirmed though, but things are much cleaner
and make more sense, so it's plausible).

Finally, the catalyst for this massive cleanup, I introduce some predefined
presets based on the data from bgo#607495.

I am not sure however how to introduce these presets. In this current
implementation, the presets are only loaded if there were no existing presets
defined by the user. Maybe it is useful to force these presets into the user's
set if the user does not have a preset existing with the same name? I need
some feedback about what a good behavior here is.