[Equalizer] support read-only presets
The default preset equalizers will now always be available, whether or not a user
has defined a preset of the same name. Default presets cannot be edited or as they
are read-only, but they will show up at the bottom of the preset list, and below
a separator, so user-defined presets are always sorted at the top of the list.

Additionally, a bug is fixed in the JSON format allowing users to have any number
of presets with the same name.