[EqualizerPresetComboBox] Fix row separator bug
Worked around what is very likely a GTK+ or Gtk# bug regarding sorting,
the row separator function, and storing null values across a row. Thanks
to Sandy Armstrong for thinking outside of his box and trying a fix
that works around the issue (causing the row separator to show up last
in the list when a new insertion is performed on the model, even though
it shows up sorted properly when the model is first constructed).

Also changed the sort function to be the default model sort, and reduced
some logic to its most simple form.