Allow debugging from MonoDevelop
Remove all custom commands from all projects options, and disable the
makefile integration for the "Execute" target. This allows MonoDevelop
to do its debugging thing with mdb. Fixes bgo#573731.

Word of caution : From now on, when starting Banshee from MonoDevelop
with "Run" or "Debug", the parameters are "--debug --uninstalled", so
your default database file will be used. I didn't find a way to pass the
value of the BANSHEE_DEV_OPTIONS env var as a parameter, which is what
"make run" does.