[Nereid] construct the player UI when first shown
Defer creating any UI until the window is actually shown for the
first time. This yields an incredible startup (~2x) startup
improvement when starting hidden (--hide). We do this in Moblin
since Banshee is a system service and is started in the session.

This allows all the services to be available, and then the Moblin
extension adds its panel to the Moblin toolbar.