New media art storage spec (bgo#520516)
This commit implements a simplified version of the proposed spec [1]

* Image files consist of a prefix ('album-', 'podcast-') and a single
MD5 hash.

 * The hash is calculated on *unmodified* artist and album name. That
is, no lower-casing, no stripping, no removal of 'special' characters
(@#!...) and no removal of what's inside brackets.

 * For album covers, the hashed string is in this format:
'<artist>\t<album>'. If we used a space, for the string "foo bar baz"
we wouldn't be able to tell whether it's album "bar baz" by "foo" or
album "baz" by "foo bar".

 * Strings are normalised to NFKD before hashing.

Thanks go to John Millikin and Bertrand Lorenz for initial work on the
spec implementation and to Gabriel Burt for overall guidance.

[1] http://live.gnome.org/MediaArtStorageSpec