[libbanshee] Fix ReplayGain support.
This is a slight reworking of the patch by Christian Taedcke on BGO# 600072.
It does not add an extra volume element; playbin2 fixes the volume change
latency on its own.

Using the rgvolume element is substantially simpler with gapless, where it's
no longer as easy to identify the start/end of a track.  This drops the custom
“use the average of the last 10 volumes as RG scaling if no RG tags” logic
Banshee currently has.