[darwin] package Monobjc as an experiment
I have no idea if I want to use this, but it looks pretty complete and
interesting. Basic testing shows some things working - GTK and the CFRunLoop
seem to integrate nicely together, and I could get volume/mount notifications
from NSWorkspace in GTK ... so that's one step closer to hardware support.

This thing was a nightmare to build and integrate though. Jeesh.