[Banshee.GStreamer] Handle slow RequestNextTrack responses better.
When the RequestNextTrack signal is not handled within 1 sec of being raised we
release the block on about-to-finish handler to let the pipeline continue.

Because state-change transitions are handled in the main thread, as is the
RequestNextTrack signal SetNextTrack will be called before the EOS transition is
handled.  In this case, store the uri sent to SetNextTrack and transition to it
in the EOS handler.