[Banshee.Metadata] Delay writing metadata for 2 seconds.
Gapless needs to update the LastPlayed field 1 sec before the track actually finishes
in order to make shuffle-by-* (and other queries which rely on LastPlayed) work.
This causes the metadata to be written out twice in quick succession.

Fix this by only writing the metadata of a track out if it has been unchanged in the
last 2 seconds.  This will have the effect of aggregating the metadata updates into
a single write.