[MeeGo] Overhaul of the panel, fix integration
This is the first drop of the re-designed MeeGo panel. This version
acts more like the main player UI, allowing the user to change sources
directly in the panel and interact with content (no longer just the
play queue). The grid view is used for displaying the album filter
if a source has one.