Allow force insert in SqliteModelProvider.Save ().
Adds a Save () overload to SqliteModelProvider which takes a second boolean
parameter to cause a forced insert.

This is needed for cases where you want to preinitialize the primary key to a
certain value. Simply calling Save () would cause the model provider to think
that this is an existing object, causing an update to happen. Result: object
not saved.

Example use case:

    private class CacheParameters
    {
        [DatabaseColumn(Constraints = DatabaseColumnConstraints.PrimaryKey)]
        public int CacheId { get; set; }

        [DatabaseColumn]
        public string Parameters { get; set; }
    }

In this class, the CacheId parameter is set to match the id of objects in
another table. Without the overload, there is no way to save it.

https://bugzilla.gnome.org/show_bug.cgi?id=613981