[bundle] bockbuild and the bundle profiles moved
This awesome bundle stuff is not Banshee specific and is useful for other
projects. For instance, I build MonoDevelop trunk with its own profile and
run it on OS X this way.

I have moved the entire bundle history from Banshee to my github:

  http://github.com/abock/bockbuild

Clone it from there, it'll be required to build Banshee bundles.