[MeeGo] we don't need to wait on MplPanelClient
My understanding of how to add contents to the MeeGo panel was flawed.
We do not need to wait on any signals (e.g. the private 'ready' signal)
to add our contents to the panel. Therefore, the windowed version and
embedded version of the panels are the same - when Banshee is ready to
add contents, BuildContents will be called and all is well.

However, after creating the MplPanelGtk instance, it seems to help to
pump the main loop. This seems to prevent mutter-moblin from saying the
panel is broken because we took too long to start up. Need to look into
that issue more deeply.