[MeeGo] add equalizer, rescan items to ToolbarMenu
Rescan is needed because the Library Watcher extension doesn't behave as
it should (doesn't do an initial scan when starting up to check for file
system differences since the app last ran).

Equalizer is added because it's nice to have on Netbooks - I was going
to integrate it directly in the toolbar, but ran out of time, so it's
in the menu for now.