[AddinView] Update model when a checkbox is toggled
Any addins that depend on the one that is disabled are disabled
automatically, so we need to update the tree view to make that visible.
We just add a call to update_model () in the Toggled handler, but that
means the code block has to be moved after the update_model declaration.