Move `Paths` to Hyena (bgo#605891)
This patch moves Paths into Hyena, which allows it to be used by other
applications.

This patch consists of a number of things:

 * The legacy application data path is moved to a new LegacyPaths class.
 * The rest of Paths is moved into Hyena.
 * All other fixes are related to getting it building with the moved
   class.

One special thing to note is the new ApplicationName member in Paths.
As I didn't want anything banshee-specific in Paths, I had to
introduce this new member. This needs to be initialized at application
startup.

Signed-off-by: Alexander Kojevnikov <alexander@kojevnikov.com>