[NowPlaying] fullscreen video is override-redirect
This patch brings improvements to fullscreen support on MeeGo/mutter,
and hopefully helps/works on other window managers. Apologies if it
doesn't, file bugs, and welcome to git master :)

With that in mind, this patch first makes the video window an
override-redirect window, so it is unknown to the window manager, and
cannot be manipulated by it (e.g. moving the video window to another
workspace in scenarios where the window manager is configured to only
place one window per workspace).

Next, to cope with what /may/ be a mutter bug, I introduce the ability
for fullscreen adapters (e.g. the X11/BaconResize adapter) to suggest
via an event that the Now Playing interface unfullscreen the video
window.

With the ability to suggest unfullscreening, the X11/BaconResize
fullscreen adapter (Banshee.NowPlaying.X11.FullscreenAdapter) now
listens to events from the window manager (via libwnck) and suggests
that the window be unfullscreened when the active workspace changes.
This is the part that works around what may be a mutter bug, where video
appears to be rendered on the new active workspace, even though the
video window (since it is override-redirect) should still be on the
previous workspace, where Banshee itself is living.

Finally, since the video window is now override-redirect, and again, the
window manager doesn't know about it, we have to make the main Banshee
window be fullscreen whenever the video window goes fullscreen, so that
the window manager knows there is actually a fullscreen window on the
workspace and can react appropriately (e.g. by not allowing toolbars or
other elements to raise above it).

When unfullscreening the video window, the fullscreen state of the main
window should be restored to whatever it was before going to video
fullscreen.

Phew.