[Audiobook] Much improved playback; add Resume action
The resume action is available if there's a last-played bookmark.
Playback can now happen from a given book (via a smart playlist limited
to that book) even while you continue to browse the rest of the
Audiobook library.  This is similar to what we do in the Music library
but there we achieve it by having child playlists that can be the
playback source - here it's an implicit source based on the playing
book.