[Metadata] Add metadata-saving support for FSQ
This patch extends SaveTrackMetadataJob to be able to support more than
one primary source to be watched for metadata changes, so we now can
disable the ReadOnly mode for the TrackEditor when the user launches it
from the FileSystemQueueSource, but without letting the file be changed
on each play if the "Save ratings and playcounts" option is enabled.
Reviewed by Alexander Kojevnikov, and some parts (like the change in
DatabaseTrackInfo.cs) done by himself as well, kudos! (BGO#609411)