Implement navigation, reloading, HTML5 spinner
Also added a loading spinner that will spin until the actual store gets
loaded. This is really exciting to me. My thought process went something
like this:

   "Okay, I'll just connect to the LoadStatusChanged event, and listen
   for when the store page begins to render; until then, I'll render
   some kind of spinner in Cairo by overriding the OssiferWebView
   OnExposeEvent virtual method; I'll of course need to use the
   animation facilities we have in Banshee to render the animation."

That probably would have been the usual 150 lines of code to put some
pixels on the screen. I'm used to this.

Then I was all like, "Whoa wait..."

  "The store is delivered using a perfectly capable WebKit HTML5
  browser, so I think this is a great excuse to write my first thingy
  using HTML5 Canvas."

And so I did. And it was easy. And it was fast. And it was pretty.