Remove bad/unnecessary AM_CONDITIONAL call
It was causing build problems for some people.

I have no idea what AM_CONDITIONAL(ENABLE_SK, false) was for, but it was
improperly implemented (since it was only called condtionally, so
ENABLE_SK may be left undefined), and nothing that I can see actually
used that conditional to begin with.