[Dap.Ipod] Save URIs when syncing to fix inconsistency (BGO#620826)
This patch does basically two things:
- Change the call to CreateTrack() by a call to CreateTrack(uri),
because for the former parameterless method it's not clear that the
assignment of the Uri property does the real copy-file operation and
that its value changes after this happens. With CreateTrack(uri), the
copy-file operation happens and the Uri property gets the location of
where the file was copied to.
- After the creation of the track, we save the Uri to banshee db (this
happens only when the track was not already on the iPod).