[Services] Refactoring of the Service addition and retrieval
This refactoring fixes the ambiguities about registering
and retrieving services, as the method Contains(string)
checks for a serviceName, while Contains<T>() checks for
the name of the class, and there could be problems if you
defined your ServiceName property with a different name
than your class. Now, after this commit:
1) When registering a service, it gets added by 2 keys
in the dictionary so you can retrieve it either way.
2) The Get<T>(string) is removed to be more in line with
the Contains overloads, and to remove clutter since T
already has the IService as a "where:" restriction and
thus can already have both ServiceName and class name.

These changes are done in ServiceManager.cs. The rest
of the changes are just to track API. A commit on b-c-e
to track API will follow this too.

Signed-off-by: Alexander Kojevnikov <alexander@kojevnikov.com>