Add initial core support for 'external' tracks
Simply put, this means tracks that are in the database and can be
searched, sorted, etc. like normal, but that will not actually play.
When an external track is activated manually, it is launched in an
external viewer.

This allows for instance, PDF files ("Digital Booklet") to be associated
with an album. An external track is currently indicated by a document
icon. More tweaks here to come.