[AudioscrobblerService] Initialize the song start time
If the scrobbling engine misses the first StartOfStream event,
for example because it was not started yet, the song start time
was not initialized. We would then try to scrobble the track with a
negative value as start time. This ended up blocking the whole
scrobbling queue.

This commit just sets a reasonable default value for the start time.
Fixes bgo#553433.