[Gio] Ensure we still have the GUdev.Device when unmounting DAPs.
When a device is unplugged and is unmounted in GIO, we can no longer
look up the corresponding GUdev.Device (which we need for various
metadata about the device). We can overcome this by storing the
GUdev.Devices in a dictionary keyed by the GIO volume. This fixes
various crashes when disconnected different devices.