Add MPRIS v2 support (bgo#570841)
This extension brings an implementation of version 2.0 of the Media
Player Remote Interfacing Specification, available at
http://www.mpris.org, allowing external applications
to interact with Banshee through a standard D-Bus API.

This API supports basic playback and media player state control, and
provides some metadata for the current track. The optional TrackList
interface is not implemented.

The extension is disabled by default for now, as there's not a lot of
MPRIS v2 clients currently in the wild. This might change in the future.

The MPRIS v2 spec make extensive use of D-Bus properties, so the
org.freedesktop.DBus.Properties interface had to be implemented
manually, including the PropertiesChanged signal.

This code is based on an old patch by John Millikin, which was an
implementation of MPRIS v1.

I'm trying to beat Aaron's record for the longest commit message, but I
think I'm not there yet, and his messages are way funnier anyway.