[Mtp] Fix various memory corruption issues
In several places we made a managed copy of a libmtp struct, then freed
the native struct, then indexed into an array on that struct. The result
was that we indexed into freed memory because the native copy did not
deep-copy the array which was on the native struct. This caused
Playlists and Albums to break. Fixes #628832.