[Dap] Add some logging for the start/end of sync processes
This adds calls to Hyena.Log.Debug when the DapSync process starts
and ends, as well as logging for the Sync Thread Cycle of the iPod
and AppleDevice extensions. Useful for measuring times in search
for performance problems, or just general helpful verbose output
that gets added to the logs that may be attached to bug reports.

Signed-off-by: Alexander Kojevnikov <alexander@kojevnikov.com>