[Dap.Ipod] Fix consistency of URIs the right way
The previous patch for BGO#620826 had some issues:
a) It still left a small window of time in which the URI would be
not the one in the device so it could still trigger wrong metadata
updates with the patch from BGO#589196.
b) It decreased performance a bit because the Save() method for
each track would trigger UI redraw.
c) Using the CreateTrack(URI) API instead of CreateTrack() and
the setter for track.Uri made ipod-sharp copy the track to the
iPod sooner (instead of at TrackDatabase.Save() time) than
before so the way to display the progress via sync_user_job
should have changed.

Now with this new approach:
* Save() isn't called until we have the correct URI, to eliminate
any period in which the track can have incorrect data in the DB.
* Save(false) is called instead of Save(true), and when the process
sync finishes, we call OnTracksAdded () and OnUserNotifyUpdated ()
to reflect the new tracks in the UI.
* A new method to display the progress during the CommitToIpod()
calls has been implemented, reusing the previous one.

Besides, this approach has already been committed for AppleDevice:
http://git.gnome.org/browse/banshee/commit/?id=c6a6aea30e56616a96872fc593ac17862ccbba79

Signed-off by: Alan McGovern <alan.mcgovern@gmail.com>