[Dap.AppleDevice] Fix the total in the progress while sync occurs
It may happen that tracks_to_add gets more songs in the middle
of a sync so we calculate the total now inside of a lock{} and
on every iteration.

This was brought up by Gabriel Burt in a review of a patch for
BGO#620826.

Signed-off by: Alan McGovern <alan.mcgovern@gmail.com>