[Dap] Infrastructure for saving metadata (Manual Sync) (bgo#589196)
This patch makes it possible for Dap sources to save metadata
changes that the user may perform after the sync has been done.
It basically adds a virtual call for the DapSource to override
and tweaks the SaveTrackMetadataJob to also take in account
tracks that were never synced before (for writing the changes
into the file too, besides the changes that the DapSource may
need to flush into its database or file system, if any).

Next patches to land will be iPod support, and AppleDevice support.
Hopefully I can have time to work on Android support later on.