[Dap.AppleDevice] Update metadata in Manual Sync mode (bgo#589196)
This patch basically takes care of implementing the UpdateMetadata()
call exposed virtually by PrimarySource to make the changes belong
to the PerformSync operation, and overrides HasEditableTrackProperties
to true for the SaveTrackMetadaService to be able to hook on the
source to flush changes to files (these changes wouldn't maybe
actually be visible by the AppleDevice but it's still good to flush
them to maintain them in sync with what is in the device's database,
just in case some user copies/imports files from the device).

Thanks goes to Alex Launi <alex.launi@gmail.com> for testing and
fixing the compilation of this patch.