[Dap.MassStorage] Add USB id for Motorola Droid X
The id is the one used for the "USB Mass Storage" mode. Ids for other
modes are as follows :
PC Mode -> 41de
Windows Media Sync -> 41d6
Those two ids are not used for now, as I'm not sure how the Droid X
behaves in those cases.

Thanks to Steve Joiner for providing the information.