[ServiceManager] Register the Network service early (bgo#627441)
This is a workaround to avoid two race conditions in ndesk-dbus :
* Initializing the Network service early means that Bus.System will be
instantiated before any other threads doing dbus stuff are started,
avoiding the first race.
* Doing it in the main thread means the code exhibiting the second race
will not be used.

See the bug for details and links to the proper fixes for those races in
dbus-sharp :
https://bugzilla.gnome.org/show_bug.cgi?id=627441

Thanks to Jérémie Laval and Alan McGovern for tracking down those race
conditions.