[GioHardware] Don't be overzealous in ignoring mount removed events
We were actually ignoring every mount removed event because when
the mount was removed, the mountpath would always be null. This worked
fine for devices who emit a VolumeRemoved event when they were
unmounted but for devices like iPhones who only emit a MountRemoved, it
meant they stayed listed in banshee forever.