[JavaScriptCore] Static function support, fixes
Static managed functions are now supported on JSClassDefinitions.
These functions are defined in managed code with a specific signature
(JSValue:JSObject,JSObject,JSValue[]) and flagged with a
JSStaticFunction attribute that specifies the JavaScript name of
the function.

Various bug fixes are made and more code is covered in the tests.

Additionally, the bridge part of the code was stubbed out a bit.