Add the SoundMenu extension from the community extensions
This extension integrates Banshee in the sound menu/indicator available
in Ubuntu.

It is disabled by default and can be enabled by passing
--enable-soundmenu to configure.

It depends on indicate-sharp to register itself with the indicator, but
this dep will probably be removed in the near future, as the
registration process is reworked.

The original source history is available in the Banshee Community
Extension repository :
http://gitorious.org/banshee-community-extensions