Add the Ubuntu One Music Store extension
This extension integrates the Ubuntu One Music Store in Banshee. It was
originally created by Jo Shields as part of the Banshee Community
Extensions.

It's disabled by default and can be enabled by passing
--enable-ubuntuone to configure.

It depends on ubuntuone-sharp, which provides a GTK+ widget embedding
the store.

The original source history is available in the Banshee Community
Extension repository :
http://gitorious.org/banshee-community-extensions